+++ src/main/java/com/bezkoder/springjwt/payload/response/Orders/OrderResp.java
@@ -17,7 +17,7 @@ public class OrderResp {
     int shoppingSum;
     int staffSum;
     double taxSum;
-    int totalSum;
+    double totalSum;
     int salarySum;
 
 }
